+++ src/main/java/org/instituteatri/deep/infrastructure/security/CustomLogoutHandler.java
@@ -4,9 +4,9 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.instituteatri.deep.domain.token.Token;
-import org.instituteatri.deep.infrastructure.exceptions.user.TokenGenerationException;
-import org.instituteatri.deep.repositories.TokenRepository;
+import org.instituteatri.deep.model.token.Token;
+import org.instituteatri.deep.exception.user.TokenGenerationException;
+import org.instituteatri.deep.repository.TokenRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
