+++ src/main/java/org/instituteatri/deep/infrastructure/security/SecurityFilter.java
@@ -7,9 +7,9 @@ import jakarta.servlet.http.HttpServletResponse;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.instituteatri.deep.infrastructure.exceptions.user.TokenInvalidException;
-import org.instituteatri.deep.repositories.TokenRepository;
-import org.instituteatri.deep.repositories.UserRepository;
+import org.instituteatri.deep.exception.user.TokenInvalidException;
+import org.instituteatri.deep.repository.TokenRepository;
+import org.instituteatri.deep.repository.UserRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
